Address PR comments
- Fix syntax errors, broken imports, unused imports, and non-essential type annotations in the new IRL algorithms (GAIL/AIRL/AMP/NEAR/TRIRL)
- Streamline prepare_expert_data/expert_data_spec to a fixed key contract for io_callback compatibility
- Fix missing mjx.forward() after Humanoid reset
- Align frame_skip/timestep with Gymnasium v5 for all custom MJX envs
- Restructure custom_mujoco/gym/*_mjx into custom_mujoco/gym/*_v5/mjx to match the existing backend-nested layout
- Enforce two-blank-line spacing across new algorithm/env files
- Add missing README Resources sections and docs/README.md links
- Remove the unimplemented IQ-Learn entry from README.md
- Compare mjx and mujoco envs through uncommited tests. Obtains similar rollouts and PPO performance
