+++ .gitignore
@@ -14,3 +14,4 @@ experiments/*.sh
 .github/evaluate-experiments.md
 experiments/plotting/
 .claude/
+/tests/
