+++ rl_x/environments/custom_mujoco/gym/half_cheetah_v5/mjx/environment.py
@@ -5,19 +5,18 @@ from functools import partial
 import jax
 import jax.numpy as jnp
 import mujoco
-from jax.scipy.spatial.transform import Rotation
 from mujoco import mjx
 
-from rl_x.environments.custom_mujoco.gym.ant_mjx.state import State
-from rl_x.environments.custom_mujoco.gym.ant_mjx.box_space import BoxSpace
-from rl_x.environments.custom_mujoco.gym.ant_mjx.viewer import MujocoViewer
+from rl_x.environments.custom_mujoco.gym.half_cheetah_v5.mjx.state import State
+from rl_x.environments.custom_mujoco.gym.half_cheetah_v5.mjx.box_space import BoxSpace
+from rl_x.environments.custom_mujoco.gym.half_cheetah_v5.mjx.viewer import MujocoViewer
 
 
-class Ant:
+class HalfCheetah:
     def __init__(self, render, horizon=1000):
         self.horizon = horizon
 
-        xml_path = (Path(__file__).resolve().parent / "data" / "ant.xml").as_posix()
+        xml_path = (Path(__file__).resolve().parent.parent / "data" / "half_cheetah.xml").as_posix()
         self.mj_model = mujoco.MjModel.from_xml_path(xml_path)
         self.mj_model.opt.solver = mujoco.mjtSolver.mjSOL_NEWTON
         self.mj_data = mujoco.MjData(self.mj_model)
@@ -26,10 +25,7 @@ class Ant:
 
         self.nr_intermediate_steps = 5
 
-        initial_height = 0.75
-        initial_rotation_quaternion = [1.0, 0.0, 0.0, 0.0]
-        initial_joint_angles = [0.0, 0.0] * 4
-        self.initial_qpos = jnp.array([0.0, 0.0, initial_height, *initial_rotation_quaternion, *initial_joint_angles])
+        self.initial_qpos = jnp.zeros(self.mjx_model.nq)
         self.initial_qvel = jnp.zeros(self.mjx_model.nv)
 
         action_bounds = self.mj_model.actuator_ctrlrange
@@ -38,17 +34,12 @@ class Ant:
         self.single_observation_space = BoxSpace(
             low=-jnp.inf,
             high=jnp.inf,
-            shape=((self.mjx_model.nq - 2) + self.mjx_model.nv + (self.mjx_model.nbody - 1) * 6,),
+            shape=((self.mjx_model.nq - 1) + self.mjx_model.nv,),
             dtype=jnp.float32,
         )
 
         self.forward_reward_weight = 1.0
-        self.healthy_z_range = (0.2, 1.0)
-        self.terminate_when_unhealthy = True
-        self.ctrl_cost_weight = 0.5
-        self.healthy_reward = 1.0
-        self.contact_force_range = (-1.0, 1.0)
-        self.contact_cost_weight = 5e-4
+        self.ctrl_cost_weight = 0.1
         self.reset_noise_scale = 0.1
 
         self.viewer = None
@@ -84,11 +75,8 @@ class Ant:
         info = {
             "rollout/episode_return": reward,
             "rollout/episode_length": 0,
-            "env_info/global_vel_x": 0.0,
             "env_info/local_vel_x": 0.0,
-            "env_info/is_healthy": 1.0,
             "env_info/ctrl_cost": 0.0,
-            "env_info/contact_cost": 0.0,
         }
         info_episode_store = {
             "episode_return": reward,
@@ -146,7 +134,7 @@ class Ant:
 
         next_observation = self.get_observation(data)
         reward, r_info = self.get_reward(data)
-        terminated = r_info["env_info/is_healthy"] < 0.5
+        terminated = False
         truncated = state.info_episode_store["episode_length"] >= self.horizon
         done = terminated | truncated
 
@@ -178,57 +166,23 @@ class Ant:
         return jax.lax.cond(done, when_done, when_not_done, None)
 
     def get_observation(self, data):
-        position = data.qpos[2:]
+        position = data.qpos[1:]
         velocity = data.qvel[:]
-        raw_contact_forces = data.cfrc_ext
-        min_value, max_value = self.contact_force_range
-        contact_forces = jnp.clip(raw_contact_forces, min_value, max_value)
-        contact_force = contact_forces[1:].flatten()
-
         observation = jnp.nan_to_num(jnp.concatenate([
             position,
             velocity,
-            contact_force,
         ]))
         return observation
 
     def get_reward(self, data):
-        torso_height = data.qpos[2]
-        base_orientation = [data.qpos[4], data.qpos[5], data.qpos[6], data.qpos[3]]
-        inverted_rotation = Rotation.from_quat(base_orientation).inv()
-        current_global_linear_velocity = data.qvel[:3]
-        current_local_linear_velocity = inverted_rotation.apply(current_global_linear_velocity)[0]
-        forward_reward = self.forward_reward_weight * current_global_linear_velocity[0]
-
-        min_z, max_z = self.healthy_z_range
-        is_healthy = jnp.clip(
-            jnp.nan_to_num(((torso_height > min_z) & (torso_height < max_z)).astype("float32")),
-            a_min=0.0,
-            a_max=1.0,
-        )
-        healthy_reward = jax.lax.cond(
-            self.terminate_when_unhealthy,
-            lambda _: self.healthy_reward,
-            lambda _: self.healthy_reward * is_healthy,
-            operand=None,
-        )
-
+        local_lin_vel = data.qvel[0]
+        forward_reward = self.forward_reward_weight * local_lin_vel
         ctrl_cost = self.ctrl_cost_weight * jnp.sum(jnp.square(data.ctrl))
-
-        raw_contact_forces = data.cfrc_ext
-        min_value, max_value = self.contact_force_range
-        contact_forces = jnp.clip(raw_contact_forces, min_value, max_value)
-        contact_force = contact_forces[1:].flatten()
-        contact_cost = self.contact_cost_weight * jnp.sum(jnp.square(contact_force))
-
-        reward = jnp.nan_to_num(jnp.clip(forward_reward, max=1e4) + healthy_reward - ctrl_cost - contact_cost)
+        reward = jnp.nan_to_num(jnp.clip(forward_reward, max=1e4) - ctrl_cost)
 
         info = {
-            "env_info/global_vel_x": current_global_linear_velocity[0],
-            "env_info/local_vel_x": current_local_linear_velocity,
-            "env_info/is_healthy": is_healthy,
+            "env_info/local_vel_x": local_lin_vel,
             "env_info/ctrl_cost": ctrl_cost,
-            "env_info/contact_cost": contact_cost,
         }
         return reward, info
