+++ rl_x/environments/custom_mujoco/gym/hopper_v5/mjx/default_config.py
