+++ rl_x/environments/custom_mujoco/gym/hopper_v5/mjx/general_properties.py
