--- rl_x/environments/custom_mujoco/gym/humanoid_mjx/__init__.py
@@ -1,8 +1,0 @@
-from rl_x.environments.environment_manager import extract_environment_name_from_file, register_environment
-from rl_x.environments.custom_mujoco.gym.humanoid_mjx.create_env import create_train_and_eval_env
-from rl_x.environments.custom_mujoco.gym.humanoid_mjx.default_config import get_config
-from rl_x.environments.custom_mujoco.gym.humanoid_mjx.general_properties import GeneralProperties
-
-
-CUSTOM_MUJOCO_HUMANOID_MJX = extract_environment_name_from_file(__file__)
-register_environment(CUSTOM_MUJOCO_HUMANOID_MJX, get_config, create_train_and_eval_env, GeneralProperties)
