+++ rl_x/environments/custom_mujoco/gym/walker2d_v5/mjx/__init__.py
@@ -1,0 +1,8 @@
+from rl_x.environments.environment_manager import extract_environment_name_from_file, register_environment
+from rl_x.environments.custom_mujoco.gym.walker2d_v5.mjx.create_env import create_train_and_eval_env
+from rl_x.environments.custom_mujoco.gym.walker2d_v5.mjx.default_config import get_config
+from rl_x.environments.custom_mujoco.gym.walker2d_v5.mjx.general_properties import GeneralProperties
+
+
+CUSTOM_MUJOCO_WALKER2D_MJX = extract_environment_name_from_file(__file__)
+register_environment(CUSTOM_MUJOCO_WALKER2D_MJX, get_config, create_train_and_eval_env, GeneralProperties)
